revised each component to its own folder
